refactor: complete modular restructuring and fix type issues
- Refactor KanbanBoard to use modular components
- Add startDate field to Task type
- Update sample data with start dates
- Fix Task type imports across all components
- Remove old drag-and-drop implementation
- Add proper TypeScript typing for priority handling
- Ensure all components use centralized Task type
- Fix ESLint configuration issues
