+++ src/data/sample-tasks.ts
@@ -1,0 +1,54 @@
+import { Task } from "@/types/task";
+
+export const sampleTasks: Task[] = [
+  {
+    id: "1",
+    title: "Learn React",
+    description: "Study React fundamentals and hooks",
+    priority: "high",
+    assignee: "John Doe",
+    status: "todo",
+    dueDate: "2024-02-15",
+    startDate: "2024-02-01",
+  },
+  {
+    id: "2",
+    title: "Build Kanban Board",
+    description: "Create a drag and drop kanban board",
+    priority: "medium",
+    assignee: "Jane Smith",
+    status: "inprogress",
+    dueDate: "2024-02-20",
+    startDate: "2024-02-05",
+  },
+  {
+    id: "3",
+    title: "Write Tests",
+    description: "Add unit tests for components",
+    priority: "low",
+    assignee: "Bob Johnson",
+    status: "done",
+    dueDate: "2024-02-10",
+    startDate: "2024-01-25",
+  },
+  {
+    id: "4",
+    title: "Design System Implementation",
+    description: "Implement consistent design tokens and components",
+    priority: "high",
+    assignee: "Alice Brown",
+    status: "inprogress",
+    dueDate: "2024-02-18",
+    startDate: "2024-02-08",
+  },
+  {
+    id: "5",
+    title: "Documentation Update",
+    description: "Update API documentation and user guides",
+    priority: "low",
+    assignee: "Charlie Wilson",
+    status: "inprogress",
+    dueDate: "2024-02-25",
+    startDate: "2024-02-12",
+  },
+];
